Update shared event links to get content under FB's updated response.
FaceBook updated how/what information it gives when an event is shared
in a feed thus I updated the plugin accordingly. The plugin has to send
another request to FaceBook to get the event date, time and location.

Changed p.caption to be p.meta as it's more descriptive; I may change
this again.